ui and backend development

+++ src/Enterprise.Framework.Application/Features/Assets/Queries/AssetDto.cs
@@ -5,10 +5,10 @@ using Enterprise.Framework.Domain.Entities;
 
 public sealed record AssetDto : IMapFrom<Asset>
 {
-    public int Id { get; init; }
+    public long Id { get; init; }
     public string Name { get; init; } = default!;
     public string SerialNumber { get; init; } = default!;
     public DateTime PurchaseDate { get; init; }
     public string Status { get; init; } = default!;
-    public int? AssignedEmployeeId { get; init; }
+    public long? AssignedEmployeeId { get; init; }
 }
