+++ src/Enterprise.Framework.Application/Features/Documents/Commands/CreateDocumentCommand.cs
@@ -10,7 +10,7 @@ public sealed record CreateDocumentCommand : IRequest<long>
     public string ContentType { get; init; } = default!;
     public string Path { get; init; } = default!;
     public string RelatedEntityType { get; init; } = default!;
-    public int RelatedEntityId { get; init; }
+    public long RelatedEntityId { get; init; }
 }
 
 public class CreateDocumentCommandHandler : IRequestHandler<CreateDocumentCommand, long>
